services/backend/requirements.txt
fastapi==0.71.0
uvicorn==0.17.0
SQLAlchemy==1.4.27
aiosqlite==0.17.0
pydantic==1.9.0
python-multipart==0.0.5
aiofiles==0.8.0
python-jose==3.3.0
cryptography==36.0.2
bcrypt==3.2.0
passlib==1.7.4